requirements.txt
celery==5.2.7
crispy-bootstrap5==0.7
Django==4.1.7
django-ckeditor==6.5.1
django-crispy-forms==2.0
django-hitcount==1.3.5
Pillow==9.4.0
psycopg2==2.9.5
